test(fuse): cover committed source finalization
