+++ packages/fuse/mod.ts
@@ -73,6 +73,7 @@ import {
   createFuseOperationState,
 } from "./operation-wiring.ts";
 import { decodeFuseComponent, encodeFusePathSegments } from "./path-codec.ts";
+import { finalizeCommittedSourceWrite } from "./source-write-finalize.ts";
 import {
   EACCES,
   EFBIG,
@@ -209,36 +210,6 @@ export function decodeSourceWriteText(
       .decode(buffer);
   } catch {
     return undefined;
-  }
-}
-
-export type CommittedSourceFinalizeResult =
-  | { status: "completed" }
-  | { status: "failed"; warning: string; error: unknown };
-
-/**
- * Run local FUSE projection work after a source transaction has committed.
- *
- * A failure here cannot undo the receipt and therefore must not reject the
- * filesystem write. Returning it as a warning keeps the durable outcome and
- * the local projection outcome separate at the same boundary as Piece does.
- */
-export async function finalizeCommittedSourceWrite(
-  receipt: PatternUpdateReceipt,
-  finalize: () => Promise<void>,
-): Promise<CommittedSourceFinalizeResult> {
-  try {
-    await finalize();
-    return { status: "completed" };
-  } catch (error) {
-    const message = error instanceof Error ? error.message : String(error);
-    return {
-      status: "failed",
-      warning: `Source revision ${receipt.revisionId} committed as ` +
-        `cf:module/${receipt.ref.identity}#${receipt.ref.symbol}, but ` +
-        `refreshing the FUSE projection failed: ${message}`,
-      error,
-    };
   }
 }
 
@@ -1881,24 +1852,21 @@ export async function main(argv: string[] = Deno.args) {
 
         bridge.writeSourceErrorLog(writeTarget.target, "");
         markExistingReady();
-        const finalized = await finalizeCommittedSourceWrite(
-          receipt,
-          () => bridge.finalizeSourceWritePath(writeTarget.target, receipt),
-        );
+        const finalize = () =>
+          bridge.finalizeSourceWritePath(writeTarget.target, receipt);
+        const finalized = await finalizeCommittedSourceWrite(receipt, finalize);
         if (finalized.status === "failed") {
-          if (isConnectionWriteFailure(finalized.error)) {
-            // The source is already durable, but the mount should still enter
-            // degraded mode when its projection refresh discovers an outage.
-            noteWriteFailure(finalized.error);
-          }
+          const error = finalized.error;
+          // The source is already durable, but the mount should still enter
+          // degraded mode when its projection refresh discovers an outage.
+          if (isConnectionWriteFailure(error)) noteWriteFailure(error);
           console.error(`[source] ${finalized.warning}`);
           bridge.writeSourceErrorLog(writeTarget.target, finalized.warning);
         }
         reconcileCfcWritebacks("source flush post-finalize");
         markExistingFinalized();
         if (handle.version === flushVersion) {
-          handle.dirty = false;
-          handle.truncatePending = false;
+          handle.dirty = handle.truncatePending = false;
         }
         writeStats.flushed++;
         console.log(`[write-trace] flush-ok ino=${handle.ino} kind=source`);
